add: correção excluir.php

+++ semestre_2/09 - Setembro/19 - bd_php_3/exemplo_1/excluir.php
@@ -4,6 +4,11 @@
 require_once('Database.php');
 $db = new Database();
 
+if(!isset($_GET['id'])) {
+    header('location:index.php');
+    die;
+}
+
 $query = 'DELETE FROM carros WHERE id = :id';
 
 $binds = [':id' => $_GET['id']];
